requirements.txt
langchain
langgraph
langchain_google_community
langchain-openai
openai
ipython
langchain_community
langchain_mcp_adapters
mcp
mysql-connector-python
